npc_fwk: Use the glib::Properties derive
- RatingLabel
- LibraryCellRenderer
- ThumbNav

+++ niepce-main/src/niepce/ui/image_grid_view.rs
@@ -66,7 +66,7 @@ impl ImageGridView {
             let renderer = item.child().and_downcast::<LibraryCellRenderer>().unwrap();
             renderer.set_property("pixbuf", image_item.thumbnail());
             renderer.set_property("libfile", image_item.file());
-            renderer.set_property("status", image_item.status() as i32);
+            renderer.set_property("status", image_item.status());
         });
 
         // Context menu
